Moved main type defs and resolvers to new dedicated files, just to store them, and added the prefix "user" to the files "user/resolvers.ts" and "user/type-defs.ts", since it was getting a little confusing when two "resolvers.ts" were open in VSCode
