+++ src/graphql/user/user-resolvers.ts
@@ -1,6 +1,8 @@
 export default {
 	Query: {
-		user: async (_parent: any, _args: any, _context: any): Promise<any> => ({}),
+		user: async (_parent: any, args: any, context: any): Promise<any> => {
+			return context.dataSources.userRepo.getItem(args.id);
+		},
 	},
 	Mutation: {
 		createUser: async (_parent: any, _args: any, _context: any): Promise<any> => ({}),
